BrineCo2Pvt: move implementation to impl file
